moved some things for citisens
